fix part of linting issues and what is to analysed

+++ ros2/src/hardware/interface/march_hardware/src/torque_sensor/torque_sensor.cpp
@@ -11,7 +11,7 @@ TorqueSensor::TorqueSensor(MotorControllerType motor_controller_type, float max_
 {
 }
 
-bool TorqueSensor::exceedsMaxTorque(float torque)
+bool TorqueSensor::exceedsMaxTorque(double torque)
 {
     return torque > max_torque_;
 }
